fix: 채팅 polling 버그 수정
