+++ back-chat/back-chat/config/routes.rb
@@ -3,8 +3,8 @@ Rails.application.routes.draw do
   get 'api/chat/list' => 'chats#list'
   get 'api/chat/join/:chat_id/:message_id' => 'chats#join'
   post 'api/chat/connect/:user_id' => 'chats#connect'
-  get 'api/chat/poll/list' => 'chats#short_poll'
-  get 'api/chat/poll/:chat_id/:message_id' => 'chats#detail_poll'
+  get 'api/chat/poll/list' => 'chats_polling#short_poll'
+  get 'api/chat/poll/:chat_id/:message_id' => 'chats_polling#detail_poll'
   post 'api/chat/:chat_id' => 'chats#send_message'
   get 'api/chat/test' => 'chats#test'
 end
