+++ back-chat/back-chat/test/controllers/chats_polling_controller_test.rb
@@ -1,0 +1,8 @@
+require "test_helper"
+
+class ChatsPollingControllerTest < ActionDispatch::IntegrationTest
+  test "should get create" do
+    get chats_polling_create_url
+    assert_response :success
+  end
+end
